+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # v1.7.0-rc.17
 ## mm/dd/2020
 
+1. [](#improved)
+    * Improve Media classes
 1. [](#bugfix)
     * Fixed `Security::sanitizeSVG()` creating an empty file if SVG file cannot be parsed
     * Fixed infinite loop in blueprints with `extend@` to a parent stream